Enable usage of date validation with presence validator
Without this change, when the user entered an invalid date, both the invalid date validation error and the date presence error would be displayed. This was happening because the date attribute writer sets the attribute value to nil if the date is invalid, as active record will not accept an invalid date and will raise an error.

It was also necessary to ensure that the review date cannot be in the past validator can handle a nil value