Fix replacement IDs updates and run metadata worker regardless of asset variant
Now that the asset model is in place, the desired behaviour for
replacement IDs is slightly different. If the update fires before
assets have become available (meaning: before the worker has finished
running and assets are saved in the WH database), we prefer not to run
the update at all. The check inside the metadata worker now uses the
`all_asset_variants_uploaded` method instead of a timestamp check.
The timestamp will be eventually removed since the presence of assets
should suffice.

To ensure that the update will fire at least once, we now call enqueue it
for every asset creation, within the `CreateAssetWorker`. Thus, on the last
asset variant creation, when all assets are available, the update will fire
for the entire AttachmentData and loop through each asset.

Co-authored-by: Tuomas Nylund tuomas.nylund@digital.cabinet-office.gov.uk