Remove placeholder logic
We were previously sending some content items to Publishing API with a
schema type prefixed `placeholder_` to indicate that content item was
still rendered by Whitehall.

This is no longer the case, so we can remove that logic.

The `GenericEditionPresenter` is used solely in tests, reverting this to
use a generic `placeholder` schema as the logic will be removed from
Publishing API in https://github.com/alphagov/publishing-api/pull/2504.

This will need to be updated again when the `placeholder` schema is
completely removed.