Use fields_for method in consultations and call for evidence form partials
We're having to manually add hidden fields for the participations,
response_form & response_form_data forms.

We can simplify this and remove a bunch of the manually created id fields
by using the fields_for method. This by default passes the id of the
forms object through to the controller.