Replace relative date validator with built-in Rails validation
Turns out we can pass a Proc to the inclusion validator that will be evaluated at runtime, making it possible to validate against a range. Therefore, our relative date validator is no longer required, we can use Rails built-in validation instead. Hurrah!

Also required an update to the validation message for the edition publication date, because we have to combine both limits into a single message.