Ensure file errors link correctly for file input
Now that there is a presence validation checking file data is present
for a consultation_response_form and a calls_for_evidence_response_form
we need to ensure that errors link correctly when the validation is
triggered.

Validations can occur for this input in 2 cases.

1. the file is invalid
2. the response_form_data is not present

This ensures that errors link correctly in both situations.