Restore default access_and_opening_time for worldwide offices
When we removed default_access_and_opening_times from worlwide orgs, some
offices which didn't have a custom access_and_opening_time weren't backfilled
with the default value that lived on the worldwide organisation.

This means thats a small subset (42) worldwide offices have an
access_and_opening_time of nil, when the worldwide organisation had a
default.

Fortunately, none of these offices have subsequently been updated post
that PR being merged.

We can backfill these with the correct value by retrieving the WorldwideOffice
from Publishing API and if the access_and_opening_time is present/different
on the object to the one we have in our database then we know that it is
one that we missed in the backfill.

We can be confident of this for 2 reasons:

1. On save the object is republished
2. No updates have occured since the PR was merged for any of these offices.

The PR was merged at 8am on 23/8/2023 so that's the date i've chosen to use here.
It will also mean if for some reason this is run in the future all, offices that
have been updated since then won't be affected.