Use GOVUK publishing date input for statistical announcements
User research indicates that teh GOVUK Publishing date input provides a superior user experience to the Rails date selection component, so we are replacing it across Whitehall.

I also removed the extracted release dates partial, as the abstraction did not survive switching components very well at all, there are too many variations on the fields name/id/parameter access required. I chose not to fix the issue where the precision input loses state if the form validation fails, as that is presently an issue in production and the fix wasn't obvious.