Give unpublished documents a new removed state
This is to make the Whitehall UI more intuitive and the state of draft more meaningful to users. It should also reduce the number of errors that occur with syncing content with the Publishing API.

From now on, users will need to click the "Create new edition" button now on unpublished documents to create a new draft, rather than the document being reset to a draft state when it was unpublished. The motive for this is that users found the previous behaviour unintuitive and they ended up with unwanted draft documents hanging around. It was also difficult to reason about for developers and caused several subtle bugs or hard to diagnose bug reports caused by something about the draft changing since unpublishing, e.g. the redirect looking like it was missing because the document slug had changed since unpublishing and therefore the same URL in production resulted in a 404.

As a result of this, we no longer need to send a new draft to the Publishing API when unpublishing a document (because no draft edition has been created)