Remove current user test
This test was added just to confirm that the Current model in Rails could be used in a thread safe way. We have been using Current without seeing any regressions so this test can probably be safely removed, as it's really just testing Rails rather than any of our code.