Run parser_errors linter and exclude two files
There's some bizzare behaviour going on in both of the excluded files.

The autocomplete.js.erb is complaining that a space is needed in
within a loop where the syntax is definitely correct.

I don't think it's worth digging through the massive file to figure out
what the issue is though as this will be removed when Bootstrap is
removed. I'm not even sure it's actually being used anywhere at the
moment.

For app/views/admin/find_in_admin_bookmarklet/_bookmarklet.erb. When you
ran the space_within_tag linter it tries to rewrite the whole function
(and breaks the functionality).

Again this is a giant file, which isn't ever touched so i dont think it's
worth the time rewriting it.