Clean up use of use_non_legacy_endpoints for AttachmentData
The use_non_legacy_endpoints flag was introduced to toggle the logic around
creating assets for whitehall attachments and using stored asset manager IDs
as the primary way of communicating with asset-manager and publishing to publishing-api,
as opposed to using legacy_url_path as the identifier.

This commit removes all the usages of the toggle for AttachmentData/file
attachments. The toggle remains in place for other data types (e.g. ImageData),
whilst implementation and migrations for those data types are pending.

The factories for attachments and attachment_data have been updated to generate
associated assets as a default.

Co-authored-by: Iris Lau iris.lau@digital.cabinet-office.gov.uk