Fix bug - excluded nations on calls for evidence
This fixes an issue reported by a user in Zendesk ticket [5494745][1],
where they're unable to set or change national exclusions on call for
evidence documents.

This is because the editions controller needs to know the names of
document types that can have national exclusions. Hardcoding the
controller names here is a bit of a smell, as pointed out in the
description of PR #6872 which introduced it.

I'd like to think there's a cleaner way to do this – I think it's ripe
for refactoring. But for now this is the quick fix.

[1]: https://govuk.zendesk.com/agent/tickets/5494745