Allow users to create new edition and edit removal explanation for removed documents
Now that we have a "removed" state for unpublished documents, users need to be able to create a new edition from them. In effect this amounts to republishing the document, but is a much simpler workflow that the current unpublishing model where the edition is moved to the "draft" state.