Remove hardcoded controller names
Following the 'quick fix' that was applied in #8295, I wanted to see if
there was anything I could do to avoid the need for hardcoding
controller names altogether.

This seems like a safe change to make here. It hasn't caused any tests
to fail. I suspect it was a hangover from the Design System Transition
work which has now all-but-concluded.

It seems reasonable just to rely on the presence of
`nation_inapplicabilities_attributes` - Edition types that don't apply
differently to different nations will not be submitting these form
fields in the first place.