Search for documents and add it to the collection group
Using the scope that is used by the edition filter in the documents search - :with_title_containing to search for live documents containing the query entered by user.
If query is empty then alert the user to enter a search query.
Link the Add button to the existing POST create_whitehall_member in the original document collection group membership controller.
This creates the membership link between the document and the collection group.
Redirect to the documents page if it is new design system.

The 'Add' button is a button rather than a link as discussed within the team and design because it is posting data, and Rails no longer supports links posting data via query params. We are collectively of the opinion that buttons should not be disguised as links, so instead of styling the button to look like a link, we should just use a secondary button.