Use correct locale for non-English edition links
The method `Edition#public_url` returns the publicly visible URL for a
document on the GOV.UK website.

However it was not returning the correct URL for editions with a
non-English primary locale – for example, for foreign-language only news
articles or consultations.

This bug was partially fixed in #7803 but only for the "View on website"
link shown in the sidebar for published editions.

We've since received a [Zendesk ticket][1] telling us that the "Preview
on website" link, as well as the links in publisher notification emails,
were still missing the locale from the end of the URL. This meant users
got "Page not found" errors when following those links.

This commit fixes the root cause of the problem, meaning I can safely
remove the fix applied in #7803 as it's no longer required.

[1]: https://govuk.zendesk.com/agent/tickets/5491481