Remove usages of uploaded_to_asset_manager timestamp
The uploaded_to_asset_manager_at field was used to track successful uploads
of assets to asset-manager. This has now been replaced by the use of the
asset model. Thus, the presence of assets in the whitehall database can be
used to ascertain whether or not the files and their variants have been
successfully uploaded to asset-manager.

Co-authored-by: Iris Lau iris.lau@digital.cabinet-office.gov.uk