Put unpublished editions into an unpublished state instead of reverting them to the draft state
This is intended to reduce confusion for users, who have reported that the behaviour does not match their expectations. The current behaviour also means that we have thousands of draft editions that we do not expect will ever be published again. This is inconsistent with other publishing apps.

In addition, the reversion to draft state makes it difficult to synchronise the content with the publishing API. Simply putting the edition into a dedicated "unpublished" state and requiring users to supersede the edition as usual makes the synchronisation logic much simpler, as we no longer need to check if the draft edition was once in the published state or send a new draft to the publishing API when we unpublish an edition.