Only render attachments in previews when we allow inline attachments
Some of our documents that do not allow inline attachments have been
wrongly showing inline attachments in the body preview on the edit
edition page. This has led to publishers being confused at where their
embedded attachment goes when they publish the document.

This was caused because we were passing the attachment ids to the
component that renders the preview based on whether they
`allow_attachments?` However publications are a document_type that
`allow_attachments?` but doesn't `allows_inline_attachments?`. And so
incorrectly shows inline attachments in the preview.

This fixes this issue by only passing the attachment ids through to the
component that renders the preview if the edition `allows_inline_attachments?`