Show processing label for images that are still uploading
If any of the image variants are still uploading, a "Processing" label will
be shown. Similar behaviour already existed for files (will be changed from
saying "Uploading" to also saying "Processing").

The behaviour for files has also been recently changed so that the label shows
when any file versions are still being uploaded/processed rather than it
showing only when the original files is missing. This behaviour is now
consistent between files and images.