Block publishing when image assets are still uploading
Extending the functionality already in place for file attachments.

If any of the 7 variants of an image (1 original + 6 carrierwave versions)
are still uploading, then the user will see an error message informing them
that they cannot publish/force-publish/submit.

This check is based on a method implemented in the ImageData class, which
verifies all existing assets associated with it.