Add deletion checks in worker
When there are any issues with the assets (i.e. the user sees a "PROCESSING" label
on the UI), the user would delete the corresponding attachment (file or image).

If the assetable type is "ImageData" the worker will retry but then subsequently
fail because the asset model has validation of presence for the assetable field.
ImageDatas do get deleted so the assetable cannot be found.

We have added a check to ensure there is no further execution if assetable has been
deleted. Added additional tests to cover for some worker retry behaviour.