Refactor the document republishing test
I've spent some time trying to understand the
PublishingApiDocumentRepublishingWorker worker – what it does, and how
it does it. I started reading the unit test but found it quite hard to
follow. I also spotted that there was some missing test coverage for a
specific code path.

I've refactored this test file with the intention of:

- improving the readability of the tests, so it's easier for future
  developers to work on it,
- testing & confirming my own understanding of the worker's expected
  behaviour, and
- adding test coverage to the `there_is_only_a_draft?` path which was
  previously untested.

It's important that the live and draft editions are sent to the
Publishing API in the correct order. There was a test for this, although
it did things in a slightly unusual way. It stubbed the Publishing API
to raise custom exceptions, which different depending on the type of
edition passed in. It then asserted that the correct type of exception
was returned, indicating that the live edition was sent first.

Instead, I've reimplemented that assertion using the [sequence][1]
matcher provided by mocha. This approach is more in-keeping with the way
sequences are usually asserted in Whitehall tests. It also means I was
able to roll the test into the main "when the document has a published
edition and a draft edition" scenario.

To help with structuring the tests, I've switched to using the MiniTest
'spec' DSL. This allowed me to clearly define the contexts that I was
testing, which would have otherwise led to long or wordy test names.

[1]: https://mocha.jamesmead.org/Mocha/API.html#sequence-instance_method