Moving preview path to asset domain, to try and fix the csv preview issue
Mission is to be able to move it to relative path so nightly data copy
dont copy the wrong hard coded production urls to lower environment.
But moving to relative path is fairly complicated and requires adding
logic to allow routing from "gov.uk" domain to asset domain with
"/media" path.
To allow fixing of some of csv's not working in prod. We are moving things
to hard-coded asset path.(This is also already the case for other attachment types).