Ensure first published inputs don't retain values from params
At the moment, when a user unchecks the previously published checkbox and
doesn't clear the date inputs, on an unsuccessful save the date inputs
retain the values because they're still present in the params.

This fixes this by pulling the values from the edition_params which
have been cleaned of the values in before actions.

Instead of using #try to check for the presence of the edition params,
I could have added an additional before_action which cleaned the params
as well as the edition params, but i'm a bit loathe to add even more
bloat to the controller.