Spike out adding custom error messages in the error summary component
This is a spike commit to test the feasibility of adding custom error
messages for specific models/attributes to locale files, while
retaining the existing behaviour for those without custom messages.

This is made slightly more difficult by presence of all the nested
error messages that can exist. For example, in this PR govspeak_content
error messages are present on the attachment object.

I've got round this for now by checking if the error is a nested error
and if that is the case is splits the string on "." and gets the second
to last element in the array.

I've had to do this rather than grab the first element in the array
because sometimes we have multiple layers of nesting.

For example, the attribute for a nested error for review reminders would be:

"document.review_reminder.attribute"

This would mean it would grab "review_reminder" and use that as the
model name it looks for in the locales files.