Apply republishing logic for new "unpublished" state
The republishing worker is used in many instances to resubmit content items to the Publishing API. Due to the introduction of the republishing state, we needed to change the way that the republishing worker handles unpublished documents.

The new behaviour is as follows:

- Resend the unpublishing data to the Publishing API. This ensures the unpublishing explanation will be updated.
- Send a draft to the publishing API if a draft exists for the document

There is no longer a need to resend the full content item, because unpublished documents are not rendered in the frontend application.