Allow organisation to update asset
Whilst other data types on which carrierwave is mounted (such as
AttachmentData and ImageData) get a new instance on update (new
file uploads), thus working as intermediaries between the document and
their attachments, organisation is itself the class that updates and
has the assets relationship.

We have added a first_or_initialize only for Organisation and kept
the previous behaviour for AttachmentData and ImageData. Organisation
cache will also be dropped in future commit if file is present to ensure
there are no overrides of the correct file.

Co-authored-by: Syed Ali <syed.ali@digital.cabinet-office.gov.uk>
Co-authored-by: Iris Lau <iris.lau@digital.cabinet-office.gov.uk>