Discard cache when file is present
The cached file will also trigger a carrierwave flow.

Since we are now allowing asset updates in the CreateAssetWorker,
chances are the cache call will run after and replace the original
file, causing a mismatch between what the Asset model says the asset
should be and the one stored in the Organisation object.

Dropping the cached file when file is present is in line with how
this issue has been addressed elsewhere, e.g. in AttachmentData.

Co-authored-by: Syed Ali <syed.ali@digital.cabinet-office.gov.uk>
Co-authored-by: Iris Lau <iris.lau@digital.cabinet-office.gov.uk>