Publish draft WW org when the state of the about CIP is submitted
Worldwide organisation "about us" corporate information pages are unusual in
that they are surfaced in the body of their owning worldwide organisation.

We previously fixed an issue where the draft page for worldwide organisations
wasn't updating when a draft about us page was created (#8058).

We've now found that any subsequent changes that republish the worldwide
organisation, do so without it's draft content.

This includes: changes to the worldwide organisation (such as a name change),
changes that republish the worldwide organisation in callbacks, and changes to
the about CIP which change it's "draft" state.

The immediate problem we want to fix is when the about CIP is submitted for 2nd
eyes. We plan to find a better solution to the wider problem later.