Filter out logos with missing assets
Logos that may have missing assets are filtered out of the presenter
load. For now we are only requiring original only as it is the only variant
defined in the uploader, but the check is future proof in case the uploader
will ever require more versions.

We have implemented an all_asset_variants_uploaded? method matching the
logic from other classes that get the mount_uploader, such as AttachmentData
and ImageData.