Delete assets when Asset is deleted from Asset Manager
Except for AttachmentData because they are soft deleted in Whitehall and Asset Manager.

Images and logos are considered less valuable resources than File attachments. They therefore
do not carry the same level of traceability and recoverability requirements as Attachments.

Whitehall deletes Image and Logos from the Asset table and considers the asset data mutable in Whitehall.

Co-authored-by: Syed Ali <syed.ali@digital.cabinet-office.gov.uk>
Co-authored-by: Laura Ghiorghisor <laura.ghiorghisor@digital.cabinet-office.gov.uk>