Add foreign key association to edition for lead images
We're going to add the ability for publishers to select a lead image
for specific case studies, and maybe down the line speeches.

This adds lead_image_id as a foreign key between the editions and
images table.

Long term, i do wonder whether it makes more sense to make this a
polymorpic association, as certian types of editions have images for
people and brand orgs as the lead image by default.

Adding it soley for editions seems like a reasonable first step for now though as
there's a lot of complexity with edition types which would need to be
ironed out first.