Handle lead_image association being set for case studies
This adds the update_lead_image method to the case study model which
handles updating the assoication when the following is true:

1. lead_image is currently nil
2. images are present
3. the user hasn't selected that they don't want a lead image

In this situation we select the image that was created first by the user
and set that as the lead image.

This method is called in 3 situations.

1. when a new image is created
2. when an image has been destroyed
3. when the image_display_option attr has been updated

This ensures we retain the exisiting behaviour of the first image that
has been added becoming the lead_image. And that when a lead image is
deleted, the case study retains a custom lead image.

Unfortunately, we need to skip validating the edition here because a
user should be able to add an image when the edition doesn't have a
change note etc.