Handle lead_image association being set on image create and delete
This adds the update_lead_image to the Edition::FirstImagePulledOut module
and calls it in 2 situations.

1. when a new image is created
2. when an image has been destroyed

The method assigns a new lead image if:

1. lead_image is currently nil
2. images are present on the edition

In this situation we select the image that was created first by the user
and set that as the lead image.

This ensures we retain the exisiting behaviour of a lead image always
being present provided an image exists (previously it used images.first).