Fix link for Document Collection Add Document by Title
When searching by title/slug to add a Whitehall page to a collection, if
there are no results from the search a link is displayed to the add by
URL page (for adding non-Whitehall pages). This link was a placeholder
until the add by URL page had been implemented—now that it has been,
set the link correctly.

Co-Authored-By: Rodney Johnson <rodney.johnson@digital.cabinet-office.gov.uk>
Co-Authored-By: Mark Taylor <mark.taylor1@digital.cabinet-office.gov.uk>