Stop marking attachments as present at unpublished
Now that we have an unpublished state for editions, we no longer need to keep track of whether attachments were present at the edition's unpublishing, because unpublished editions have to be superseded. Previously unpublished editions were put into the draft state and it was therefore necessary to track which attachments were present when it was unpublished so that we could distinguish them from attachments which had been added to the draft since unpublishing.