Handle lead_image association being set for news articles
This adds the update_lead_image method to the news articles model. It is
similiar to the previous commit which set lead_image_id for case studies
and handles setting the lead_image_id to the id of the oldest image
when an image is created and the lead_image_id is null. Or if
the lead_image is deleted.