Ensure lead_image_id is correct for new draft editions
When a user creates a new edition, we duplicate the assoications via
the Edition#create_draft method.

We need to ensure that the when the images are duplicated the editions
lead_image_id is updated to the newly duped instance of the image
that was previously the lead_image.

Without doing this the edition would continue to use the lead_image from
the live edition.