Fix unpublished edition attachment redirects when draft edition exists
This fixes a bug introduced in 4eb10f4. In that commit we stopped preventing users from unpublishing a document if a draft edition exists. However, the attachment redirect URL update behaviour assumed that the latest attachable would always be the unpublished one and failed to update the redirect if it found a draft.

The new behaviour checks for the existence of an unpublished attachable and updates the redirect if it finds one. There can only ever be a single unpublished attachment at the moment, so this is a safe approach.