Update reporting rake task to include all govspeak columns
This adds all the fields which could contain govspeak into the
`reporting:matching_docs` rake task, allowing us to see all document
types that would need to be republished if part of the govspeak gem were
to be updated.

Replacing the existing use of `find_in_batches` and then `each` with a
single use of `find_each`, which does the same thing (with a default
batch size of 1,000).