Put attachment redirect and replacement responsibilities into separate methods on AttachmentUpdater
Previously, the AttachmentUpdater was using a single method to perform three distinct operations: update the attachment status, update the attachment redirect for unpublished documents, and replacing the attachment data assets. By using separate methods for each operation, we reduce a lot of the AttachmentUpdater's complexity. We no longer have to provide flags for which fields we want to be passed to the AssetUpdater, and we no longer need to build a hash and then strip out any nil values before passing it to the AssetUpdater. In addition, this commit moves some of the AttachmentUpdater logic onto the AttachmentData model which provides a richer domain model and a better separation of responsibilities.