Update Topical Event logo model in place when the file is updated
When logo file is replaced, the previous behaviour would not delete assets from asset manager.
This change makes it so that the Featured Image Data instance is reused for the new file and the related assets in asset manager are deleted.

It looks like the previous behaviour was in place because of a possible time delay between Whitehall publishing a document and
a dependency resolution process in Publishing Api checking that any dependent pages are updated as well.

Now time has passed and knowledge whether this is any longer a problem or not has disappeared with the people who have left GDS.
It looks like we could change this approach, therefore this commit will change it. If any broken pages appear due to the dependency resolution
running later than the deletion of the previous image - then we will revert this change and keep the previous image alive in asset manager as well.

Co-authored-by: Laura Ghiorghisor <laura.ghiorghisor@digital.cabinet-office.gov.uk>