Ensure update_image_display_option updates lead image
This does 3 things:

1. moves the update_image_display_option endpoint to the case studies
controller. Only case studies use the image_display_option
attribute to toggle using a custom image, organisation image or
no image so it should probably live in that controller
2. calls #update_lead_image on a successfull save
3. updates #update_lead_image to delete the edition_lead_image
if #update_lead_image if the edition has been updated to have no_image
or to show the organisation image.