Add edition_lead_image joins table and model associations
We're going to add in functionality for users to select a lead image.

This creates a lead_image table, which contains and image_id and edition_id
and adds the relevant associations in the models.

This should be a has_one as an edition should only ever be able to have
one lead image. I've added a unique constraint to on edition_id to
enforce this.