Publish document types like Organisation and Topical event when assets are created
Some document types are not editioned/versioned and they need to be published immediately when
the document changes. For example Organisation title change goes live immediately.

Assets are created in an asynchronous manner and therefore the Document needs to be published
again when the last asset variant is saved. This allows PublishingApiPresenter to send the images
to PublishingApi.

Co-authored-by: syed-ali-tw <syed.ali@digital.cabinet-office.gov.uk>