Fix bug with update_lead_image method
The 'and return if' isn't returning when the image_display_option is
no_image or organisation_image.

This means that when an image is created and the lead_image isn't set
the assoication is created.

It still wouldn't be pushed downstream to Publishing API, but we
shouldn't have lead images sitting in the db when there isn't a
lead image.

This updates the method to explicitly call return within the conditional
which resolves the issue.