Fix orphaned attachment data error
Fixes bug introduced in 78eb085. That commit added a guard clause which was intended to prevent attachments that had become unattached from their edition from being updated in asset manager, but that was based on an incorrect assumption about how the system should behave. As far as we're aware, there shouldn't be attachments without an attachable.

However, we do need to ensure that attachment data records that have been replaced and are therefore no longer associated with any attachments can be updated. This commit adds a test to ensure that the behaviour matches that of the AttachmentUpdater prior to 78eb085. We're not entirely convinced that the behaviour is logical and we intend to do further research to understand it better.

Co-authored-by: Tuomas Nylund <tuomas.nylund@thoughtworks.com>
Co-authored-by: Laura Ghiorghisor <laura.ghiorghisor@digital.cabinet-office.gov.uk>