Installed FlipFlop gem
We are installing the FlipFlop gem to enable us to use feature flags with a lower barrier to entry that the current process of using user permissions. The FlipFlop gem can store feature flag information inside the Whitehall database, meaning that we no longer have to create permissions inside of signon to deploy code without releasing it to users.

We have disabled the cookie strategy because our main use case at the moment is toggling features on and off in different environments, but we may consider using it in future if our needs change.

We have also only enabled the Flipflop user interface in local development environments. Developers will use the rake task provided by the library to toggle features in cloud environments.

This commit also switches the FlipFlop strategy to use an in-memory store for feature flags and makes the flag data available to test cases.