Display default organisation image when it's being used
There's some pretty gnarly behaviour going on here under the hood.

The NewsArticlePresenter handles passing an image through to Publishing API.

It calls LeadImage#image_data which checks for a lead image by:

- checking if a custom lead image is present
- then checking if the first lead org has a default news image
- then checking if the first org has a default news image
- then checks if owned by a worldwide org, if so it passes their default
news image

If no image is found is passes an empty hash to to the details blob.
Unforunately, government-frontend overides this and renders a placeholder
image if none is passed, so we still need to account for that.

Case studies behave differently. It doesn't use the LeadImage#image_data
method and instead:

- checks if a custom lead image is present
- then checks if the first lead org has a default news image

If not image is found then it passes an empty hash which is respected by
the frontend application.